Get-KnownFolder: facilitate obtaining common folder sets
Add switches specifying which folder sets will be returned.
-peruser: all KnownFolders shown in the user profile folder by Explorer,
          except for SkyDrive
-public: all folders shown in the public profile by Explorer
-all: all KnownFolders registered in the system

If no arguments are specified, -peruser is assumed.

The intention is to simplify the common case of redirecting all folders
with valuable user content (in contrast to application-created data -
AppData) to another drive/partition. Rule of thumb: this is the data that
needs to survive OS reinstall. As SkyDrive is simply a local cache of
cloud-stored data, it is also excluded.